graphql/core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "syntek-graphql-core"
version = "1.0.0"
description = "Security foundation for Syntek GraphQL modules - provides core security components, error handling, and utilities"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.14"
authors = [
    { name = "Syntek Development Team", email = "dev@syntek.dev" },
]
keywords = [
    "graphql",
    "security",
    "django",
    "strawberry",
    "authentication",
    "authorization",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: Django :: 6.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "django>=6.0.2",
    "strawberry-graphql>=0.291.0",
    "redis>=5.2.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-django>=4.10.0",
    "pytest-cov>=6.0.0",
    "mypy>=1.15.0",
    "ruff>=0.9.1",
]

[project.urls]
Homepage = "https://github.com/syntek/syntek-modules"
Documentation = "https://docs.syntek.dev/graphql/core"
Repository = "https://github.com/syntek/syntek-modules"
Issues = "https://github.com/syntek/syntek-modules/issues"

[tool.hatch.build.targets.wheel]
packages = ["syntek_graphql_core"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
strict_optional = true
plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
module = "strawberry.*"
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "tests.settings"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.coverage.run]
source = ["syntek_graphql_core"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
]
